+++ CHANGELOG.md
@@ -49,6 +49,7 @@ FIXME: Released on
 - Bugfix:
   - SCP client didn't show file types for files
   - FTP client didn't show file types for files
+  - Fixed `0 B/S` transfer rate displayed after completing download in less than 1 second
 - Dependencies:
   - added `bitflags 1.2.1`
   - removed `data-encoding`